# apps/web/api.yaml
openapi: 3.0.0
paths:
...
  /api/vendedores:
    get:
      description: Retorna uma lista de todos os usuários com a role "VENDEDOR".
      operationId: VendedorController_findAll
      parameters: []
      responses:
        '200':
          description: Lista de vendedores retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VendedorResponseDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      security:
        - bearer: []
      summary: Lista todos os vendedores
      tags:
        - Vendedores
  /api/vendedores/{vendedorId}/indicacoes:
    post:
      operationId: VendedorController_criarIndicacao
      parameters:
        - name: vendedorId
          required: true
          in: path
          description: UUID do vendedor
          schema:
            example: a1b2c3d4-e5f6-7890-1234-567890abcdef
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndicacaoDto'
      responses:
        '204':
          description: ''
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      security:
        - bearer: []
      summary: Cria um novo vínculo (indicação) entre um vendedor e um estabelecimento.
      tags:
        - Vendedores
    get:
      operationId: VendedorController_listarIndicacoesPorVendedor
      parameters:
        - name: vendedorId
          required: true
          in: path
          description: UUID do vendedor
          schema:
            example: a1b2c3d4-e5f6-7890-1234-567890abcdef
            type: string
      responses:
        '200':
          description: Lista de indicações retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IndicacaoResponseDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      security:
        - bearer: []
      summary: Lista todos os estabelecimentos indicados por um vendedor.
      tags:
        - Vendedores
  /api/vendedores/{vendedorId}/indicacoes/{estabelecimentoId}:
    patch:
      operationId: VendedorController_atualizarTaxaIndicacao
      parameters:
        - name: vendedorId
          required: true
          in: path
          description: UUID do vendedor
          schema:
            type: string
        - name: estabelecimentoId
          required: true
          in: path
          description: UUID do estabelecimento
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIndicacaoDto'
      responses:
        '204':
          description: Taxa de comissão atualizada com sucesso.
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      security:
        - bearer: []
      summary: Atualiza a taxa de comissão de uma indicação específica.
      tags:
        - Vendedores
    delete:
      operationId: VendedorController_removerIndicacao
      parameters:
        - name: vendedorId
          required: true
          in: path
          description: UUID do vendedor
          schema:
            type: string
        - name: estabelecimentoId
          required: true
          in: path
          description: UUID do estabelecimento
          schema:
            type: string
      responses:
        '204':
          description: Vínculo removido com sucesso.
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
      security:
        - bearer: []
      summary: Remove o vínculo (indicação) entre um vendedor e um estabelecimento.
      tags:
        - Vendedores
...
info:
  title: Portal Unipag
  description: Documentação do portal unipag
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    HttpError:
      type: object
      properties:
        statusCode:
          type: number
        message:
          type: string
        error:
          type: string
      required:
        - statusCode
        - message
...
    VendedorResponseDto:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        nome:
          type: string
          example: João da Silva
        email:
          type: string
          example: joao.silva@example.com
        telefone:
          type: object
          example: '11987654321'
          nullable: true
        role:
          type: string
          enum:
            - ADMINISTRADOR
            - VENDEDOR
          example: VENDEDOR
        criado_em:
          format: date-time
          type: string
        ativo:
          type: boolean
          example: true
      required:
        - id
        - nome
        - email
        - telefone
        - role
        - criado_em
        - ativo
    CreateIndicacaoDto:
      type: object
      properties:
        estabelecimento_id:
          type: string
          description: UUID do estabelecimento que está sendo indicado.
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        taxa_comissao:
          type: number
          description: >-
            Taxa de comissão (percentual) para o vendedor sobre este
            estabelecimento.
          example: 3.75
      required:
        - estabelecimento_id
        - taxa_comissao
    EstabelecimentoResponseDto:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        ceo_pag_id:
          type: number
          example: 12345
        document_number:
          type: string
          example: 12.345.678/0001-99
        social_reason:
          type: string
          example: Empresa Exemplo LTDA
        name:
          type: string
          example: Nome Fantasia Exemplo
        status:
          type: number
          example: 1
        account_source:
          type: string
          example: ONE
        created_at:
          format: date-time
          type: string
          example: '2023-10-01T12:00:00Z'
        updated_at:
          type: object
          example: '2023-10-01T12:00:00Z'
      required:
        - id
        - ceo_pag_id
        - document_number
        - social_reason
        - name
        - status
        - account_source
        - created_at
    IndicacaoResponseDto:
      type: object
      properties:
        id:
          type: string
          example: f0e9d8c7-b6a5-4321-fedc-ba9876543210
        taxa_comissao:
          type: number
          description: Taxa de comissão percentual
          example: 5.5
        atribuido_em:
          format: date-time
          type: string
        usuario_id:
          type: string
          example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        estabelecimento_id:
          type: string
          example: b1c2d3e4-f5g6-7890-1234-567890abcdef
        estabelecimento:
          $ref: '#/components/schemas/EstabelecimentoResponseDto'
      required:
        - id
        - taxa_comissao
        - atribuido_em
        - usuario_id
        - estabelecimento_id
        - estabelecimento
    UpdateIndicacaoDto:
      type: object
      properties:
        taxa_comissao:
          type: number
          description: Nova taxa de comissão (percentual) para a indicação.
          example: 4
      required:
        - taxa_comissao
...